djangoboke/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudmusic==0.1.1
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-model-utils==4.2.0
django-mptt==0.14.0
django-notifications-hq==1.7.0
django-taggit==3.0.0
idna==3.4
importlib-metadata==5.0.0
jsonfield==3.1.0
lxml==4.9.1
Markdown==3.4.1
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
pycryptodomex==3.15.0
Pygments==2.13.0
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
swapper==1.3.0
threadpool==1.3.2
tzdata==2022.6
urllib3==1.26.12
zipp==3.10.0
